requirements.txt
aiohttp==3.13.2
fastapi==0.120.2
passlib==1.7.4
pydantic==2.12.3
pydantic_settings==2.11.0
PyJWT==2.10.1
SQLAlchemy==2.0.44
starlette==0.49.1
